==> notebooks/1_feature_extraction.ipynb <==
import os
import pandas as pd

data_directory = '../dataset/MYCN_Student_ROI_dataset'
ground_truth_file_path = '../dataset/neuroblast_GT_preprocessed.csv'

ground_truth = pd.read_csv(ground_truth_file_path)

ground_truth.head()

from radiomics import featureextractor

extractor = featureextractor.RadiomicsFeatureExtractor()

features_list = []
label_list = []

# Iterate through all the files to extract features
for sample_folder in os.listdir(data_directory):
    sample_path = os.path.join(data_directory, sample_folder)

    image_path = os.path.join(sample_path, sample_folder + '_roi_image.nii.gz')
    mask_path = os.path.join(sample_path, sample_folder + '_roi_image_gt.nii.gz')

    features = extractor.execute(image_path, mask_path)
    features_list.append(features) # List of feature to value dictionary

    label = ground_truth.loc[ground_truth['sample_id'] == sample_folder].values[0]
    label_list.append(label) # List of id to label dictionary 

features_df = pd.DataFrame(features_list)
label_df = pd.DataFrame(label_list, columns=['sample_id', 'mycn_amplified'])
display(features_df.head())
display(label_df.head())
print('Shape of features_df:', features_df.shape)
print('Shape of label_df:', label_df.shape)

if features_df.isna().any().any() == True:
    print('There are missing values')
elif features_df.isna().any().any() == False:
    print('There are no missing values')

df = pd.concat([features_df, label_df], axis=1, join='outer')
display(df.head())
print(df.shape)

samples_post_extraction_and_labelled_df = df.loc[final_df['mycn_amplified'] != -1]
display(samples_post_extraction_and_labelled_df.head())
print(samples_post_extraction_and_labelled_df.shape)
samples_post_extraction_and_labelled_df.to_csv('../dataset/samples_post_extraction_and_labelled.csv', index=False)
